AB#45735 Moved functions for Week and Years to NMDataService from Controller
