+++ UKHO.MaritimeSafetyInformation.Web/UKHO.MaritimeSafetyInformation.Web/Services/Interfaces/IFileShareService.cs
@@ -6,6 +6,5 @@ namespace UKHO.MaritimeSafetyInformation.Web.Services.Interfaces
     public interface IFileShareService
     {
         public Task<IResult<BatchSearchResponse>> FssWeeklySearchAsync(string searchText,string accessToken);
-        public Task<Stream> GetFssFileStreamAsync(string batchId, string filename, string accessToken);
     }
 }
