+++ UKHO.MaritimeSafetyInformation.Web/UKHO.MaritimeSafetyInformation.Web/Services/Interfaces/INMDataService.cs
@@ -5,7 +5,8 @@ namespace UKHO.MaritimeSafetyInformation.Web.Services.Interfaces
 {
     public interface INMDataService
     {
-        public Task<List<ShowFilesResponseModel>> GetBatchDetailsFiles(int year, int week, string accessToken);
-        public Task<byte[]> DownloadFssFileAsync(string batchId, string filename, string accessToken);
+        public Task<List<ShowFilesResponseModel>> GetBatchDetailsFiles(int year, int week);
+        public List<KeyValuePair<string, string>> GetPastYears();
+        public List<KeyValuePair<string, string>> GetAllWeeksofYear(int year);
     }
 }
