+++ UKHO.MaritimeSafetyInformation.Web/UKHO.MaritimeSafetyInformation.Web/Views/NoticestoMariners/ShowWeeklyFiles.cshtml
@@ -1,5 +1,4 @@
-﻿@using UKHO.FileShareClient.Models
-@model List<ShowFilesResponseModel>
+﻿@model List<ShowFilesResponseModel>
 @section Scripts
 {
     <script type="text/javascript" src="@Url.Content("~/js/Custom/NoticesToMariners.js")"></script>
