+++ UKHO.MaritimeSafetyInformation.Web/UKHO.MaritimeSafetyInformation.Web/Views/NoticestoMariners/_WeeklyFilesList.cshtml
@@ -1,4 +1,4 @@
-﻿@using UKHO.FileShareClient.Models
+﻿@using UKHO.MaritimeSafetyInformation.Web.Models
 @model List<ShowFilesResponseModel>
 @{
     var a = Model;
@@ -10,7 +10,7 @@
             @foreach (var item in Model)
             {
                 <tr>
-                    <td><a href="@item.Links.Get.Href">@item.FileDescription</a></td>
+                    <td><a href="#">@item.FileDescription</a></td>
                     <td>@item.FileSizeinKB (@item.FileExtension)</td>
                     <td></td>
                 </tr>
